can-example-app: fix null pointer exception when accessing socket
If an error occurs creating/binding the raw socket, the socket handle
will be null. This usually happens when permissions on the device are
insufficient to create/bind sockets in the system. If sockets cannot
be created at startup, notify the user right away by displaying a custom
view with the error.

Signed-off-by: Laura Nao <laura.nao@kynetics.com>